+++ CHANGELOG.md
@@ -146,5 +146,7 @@ significant modifications will be credited to OpenTelemetry Authors.
 ([#600](https://github.com/open-telemetry/opentelemetry-demo/pull/600))
 * Add HTTP client instrumentation to shippingservice
 ([#610](https://github.com/open-telemetry/opentelemetry-demo/pull/610))
+* Added Kafka, accountingservice and frauddetectionservice for async workflows
+([#512](https://github.com/open-telemetry/opentelemetry-demo/pull/457))
 * Add tracing to Envoy (frontend-proxy)
 ([#613](https://github.com/open-telemetry/opentelemetry-demo/pull/613))